Revert "Merge branch 'extrazoom'"
This reverts commit 3b30e0997a970810d273d321405b67b923efe4b5, reversing
changes made to 50b82826628f0273d6f88f9dd49af35bbec06fd0.

+++ src/spritecache.h
@@ -14,8 +14,9 @@
 
 #include "gfx_type.h"
 
+/** Data structure describing a sprite. */
 struct Sprite {
-	uint16 height;  ///< Height of the sprite.
+	byte height;  ///< Height of the sprite.
 	uint16 width; ///< Width of the sprite.
 	int16 x_offs; ///< Number of pixels to shift the sprite to the right.
 	int16 y_offs; ///< Number of pixels to shift the sprite downwards.
